refactor(scenario): retire YAML half; relocate helpers into test_bench
Part 1 of issue 821 (iamacoffeepot/aether#821). Internal cleanup with
zero consumer test files touched — every existing aether-scenario
import keeps compiling via shim re-exports. PR 2 will inline the
Script/Step/Check vocabulary at the four consumer call sites and
retire the crate.

Retired (zero consumers outside aether-scenario's own self-tests):

- aether-scenario-macros crate (scenario_dir! proc-macro)
- aether-scenario bin (YAML runner CLI)
- crates/aether-scenario/tests/fixtures/ (YAML scenario fixtures)
- tests/macro_scenario_dir.rs, cli_round_trip.rs, runner_round_trip.rs
- runner::run_yaml_str, script::parse_script
- RunnerError::Parse variant

Relocated into aether-substrate-bundle::test_bench:

- src/test_bench/test_helpers.rs — has_wgpu_adapter,
  locate_component_wasm, require_runtime, init_save_sandbox,
  test_namespace_roots, write_fixture (all script-vocabulary-free)
- src/test_bench/visual.rs — decode_png, differs_from_background,
  not_all_black, Image, ImageError

aether-scenario stays as a thin layer: it re-exports the relocated
helpers, keeps the Script/Step/Check/Runner Rust vocabulary, and
retains run_or_panic + tick_to (the two helpers bound to Script /
Step that retire in PR 2 with the vocabulary).

Deps shed from aether-scenario: aether-scenario-macros, png, pollster,
wgpu, thiserror was already there.

Deps added to aether-substrate-bundle: thiserror (for the moved
ImageError; the crate had no prior thiserror consumer).

Verification: cargo nextest run --workspace = 974/974 passed;
cargo clippy --all-targets -- -D warnings clean; cargo fmt -- --check
clean; cargo test --doc --workspace clean.
