+++ crates/aether-scenario/Cargo.toml
@@ -6,38 +6,28 @@ edition.workspace = true
 # ADR-0067 scenario library: declarative Script → Runner → RunReport
 # pipeline that drives an in-process `TestBench` chassis through a
 # sequence of steps and asserts visual properties of captured frames.
-# This crate ships the core types + a YAML parser; per-component
-# scenarios live in their owning crates' tests/ directories and the
-# CLI wrapper gives agents a one-shot "run scenario at path" entry.
+#
+# Issue 821 PR 1 retired the YAML half (parser, bin, scenario_dir!
+# proc-macro) and relocated `test_helpers` + `visual` into
+# `aether_substrate_bundle::test_bench`. This crate is now a thin
+# layer exposing the `Script` / `Step` / `Check` / `Runner` Rust
+# vocabulary plus shimmed re-exports of the relocated helpers.
+# Issue 821 PR 2 will inline the vocabulary at the four consumer
+# call sites and retire the crate.
 
 [lib]
 path = "src/lib.rs"
-
-# Thin CLI wrapper. Takes a path to a YAML scenario script, boots a
-# TestBench, runs the script, prints a report, exits 0 on pass / 1
-# on fail. The agent-facing entry point — component developers
-# running scenarios from the IDE use the `scenario_dir!` proc-macro
-# instead (`aether-scenario-macros`).
-[[bin]]
-name = "aether-scenario"
-path = "src/bin/aether-scenario.rs"
 
 [dependencies]
 aether-data = { path = "../aether-data" }
 aether-kinds = { path = "../aether-kinds" }
 aether-codec = { path = "../aether-codec" }
-aether-scenario-macros = { path = "../aether-scenario-macros" }
 aether-substrate-bundle = { path = "../aether-substrate-bundle" }
-png = "0.17"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
+# `Step::SendMail.params` carries a `serde_yml::Value` so consumers
+# can build the parameter tree using YAML idioms (string-keyed maps,
+# scalars, arrays). Retires in PR 2 of issue 821 along with the
+# Script vocabulary.
 serde_yml = "0.0.12"
 thiserror = "1"
-# `test_helpers::has_wgpu_adapter` probes for a wgpu adapter so per-
-# component scenario tests can skip cleanly on driverless dev boxes.
-# `pollster` blocks on the adapter request future without dragging in
-# a full async runtime. Both used to be dev-deps; promoted to runtime
-# deps so consumers can call the helpers from their own integration
-# tests via `pub mod test_helpers` (issue 460).
-pollster = "0.4.0"
-wgpu = "29.0.1"
