+++ crates/aether-scenario/src/script.rs
@@ -107,180 +107,3 @@ fn default_tolerance() -> u8 {
 fn default_min_count() -> usize {
     1
 }
-
-/// Parse a script from YAML source. Returns the underlying serde_yml
-/// error wrapped with file context handled at call sites.
-pub fn parse_script(yaml: &str) -> Result<Script, serde_yml::Error> {
-    serde_yml::from_str(yaml)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn round_trip_basic_script() {
-        let yaml = r#"
-name: empty boot
-steps:
-  - op: advance
-    ticks: 1
-  - op: capture
-  - op: assert
-    check:
-      kind: not_all_black
-"#;
-        let script = parse_script(yaml).expect("parse");
-        assert_eq!(script.name, "empty boot");
-        assert_eq!(script.steps.len(), 3);
-        match &script.steps[0] {
-            Step::Advance { ticks } => assert_eq!(*ticks, 1),
-            other => panic!("expected Advance, got {other:?}"),
-        }
-        match &script.steps[1] {
-            Step::Capture => (),
-            other => panic!("expected Capture, got {other:?}"),
-        }
-        match &script.steps[2] {
-            Step::Assert { check } => {
-                assert!(matches!(check, Check::NotAllBlack));
-            }
-            other => panic!("expected Assert, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn parses_differs_from_background_with_default_tolerance() {
-        let yaml = r#"
-name: bg
-steps:
-  - op: assert
-    check:
-      kind: differs_from_background
-"#;
-        let script = parse_script(yaml).expect("parse");
-        match &script.steps[0] {
-            Step::Assert { check } => match check {
-                Check::DiffersFromBackground { tolerance } => assert_eq!(*tolerance, 5),
-                other => panic!("expected DiffersFromBackground, got {other:?}"),
-            },
-            other => panic!("expected Assert, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn parses_mail_observed_with_default_min_count() {
-        let yaml = r#"
-name: mail
-steps:
-  - op: assert
-    check:
-      kind: mail_observed
-      name: aether.draw_triangle
-"#;
-        let script = parse_script(yaml).expect("parse");
-        match &script.steps[0] {
-            Step::Assert { check } => match check {
-                Check::MailObserved { name, min_count } => {
-                    assert_eq!(name, "aether.draw_triangle");
-                    assert_eq!(*min_count, 1);
-                }
-                other => panic!("expected MailObserved, got {other:?}"),
-            },
-            other => panic!("expected Assert, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn parses_mail_not_observed() {
-        let yaml = r#"
-name: no-mail
-steps:
-  - op: assert
-    check:
-      kind: mail_not_observed
-      name: aether.audio.note_on
-"#;
-        let script = parse_script(yaml).expect("parse");
-        match &script.steps[0] {
-            Step::Assert { check } => match check {
-                Check::MailNotObserved { name } => {
-                    assert_eq!(name, "aether.audio.note_on");
-                }
-                other => panic!("expected MailNotObserved, got {other:?}"),
-            },
-            other => panic!("expected Assert, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn empty_steps_is_valid() {
-        let script = parse_script("name: nothing\nsteps: []\n").expect("parse");
-        assert_eq!(script.name, "nothing");
-        assert!(script.steps.is_empty());
-    }
-
-    #[test]
-    fn missing_op_field_errors() {
-        let yaml = "name: bad\nsteps:\n  - ticks: 1\n";
-        assert!(parse_script(yaml).is_err());
-    }
-
-    #[test]
-    fn parses_load_component_and_send_mail_steps() {
-        let yaml = r#"
-name: drive a component
-steps:
-  - op: load_component
-    path: /tmp/mesh-viewer.wasm
-    name: mv
-  - op: send_mail
-    recipient: aether.fs
-    kind: aether.fs.write
-    params:
-      namespace: save
-      path: greeting.txt
-      bytes: [104, 105]
-"#;
-        let script = parse_script(yaml).expect("parse");
-        assert_eq!(script.steps.len(), 2);
-        match &script.steps[0] {
-            Step::LoadComponent { path, name } => {
-                assert_eq!(path, "/tmp/mesh-viewer.wasm");
-                assert_eq!(name.as_deref(), Some("mv"));
-            }
-            other => panic!("expected LoadComponent, got {other:?}"),
-        }
-        match &script.steps[1] {
-            Step::SendMail {
-                recipient,
-                kind,
-                params,
-            } => {
-                assert_eq!(recipient, "aether.fs");
-                assert_eq!(kind, "aether.fs.write");
-                let mapping = params.as_mapping().expect("params is mapping");
-                assert_eq!(
-                    mapping.get(serde_yml::Value::String("namespace".to_owned())),
-                    Some(&serde_yml::Value::String("save".to_owned()))
-                );
-            }
-            other => panic!("expected SendMail, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn load_component_name_defaults_to_none() {
-        let yaml = r#"
-name: just-load
-steps:
-  - op: load_component
-    path: /tmp/x.wasm
-"#;
-        let script = parse_script(yaml).expect("parse");
-        match &script.steps[0] {
-            Step::LoadComponent { name, .. } => assert!(name.is_none()),
-            other => panic!("expected LoadComponent, got {other:?}"),
-        }
-    }
-}
