--- crates/aether-scenario/tests/fixtures/clear_color.yml
@@ -1,8 +1,0 @@
-name: clear color cli round trip
-steps:
-  - op: advance
-    ticks: 1
-  - op: capture
-  - op: assert
-    check:
-      kind: not_all_black
