--- crates/aether-scenario/tests/fixtures/scenarios/clear_color_round_trip.yml
@@ -1,8 +1,0 @@
-name: clear color round trip
-steps:
-  - op: advance
-    ticks: 1
-  - op: capture
-  - op: assert
-    check:
-      kind: not_all_black
