--- crates/aether-scenario/tests/fixtures/scenarios/empty_boot.yml
@@ -1,2 +1,0 @@
-name: empty boot
-steps: []
