--- crates/aether-scenario/tests/fixtures/scenarios/empty_boot_observes_no_geometry.yml
@@ -1,19 +1,0 @@
-name: empty boot observes no geometry
-# An empty bench should clear to the chassis background and never
-# produce DrawTriangle. Validates mail-observation and visual asserts
-# on the baseline state. NotAllBlack should pass (background is
-# non-black); MailNotObserved for DrawTriangle should pass.
-# Pre-#775 this script also asserted MailNotObserved against
-# `aether.observation.frame_stats` at <120 ticks; the kind retired
-# alongside BroadcastCapability so the assertion went with it.
-steps:
-  - op: advance
-    ticks: 5
-  - op: capture
-  - op: assert
-    check:
-      kind: not_all_black
-  - op: assert
-    check:
-      kind: mail_not_observed
-      name: aether.draw_triangle
